[ALLUXIO-1907] second attempt to deflake free tests
